Move log subcommands from utils to log component
