+++ austrakka/components/log/__init__.py
@@ -1,0 +1,4 @@
+
+# This component has reusable subcommands which are added to multiple commands,
+# including top-level
+from .subcommands import log_subcommands
